--- ambry-clustermap/src/main/java/com/github/ambry/clustermap/CloudServiceClusterInfo.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2020 LinkedIn Corp. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *
- */
-package com.github.ambry.clustermap;
-
-import com.github.ambry.config.ClusterMapConfig;
-
-
-public class CloudServiceClusterInfo extends ClusterInfo {
-  /**
-   * Construct a DcInfo object with the given parameters.
-   * @param clusterChangeHandler the associated {@link ClusterChangeHandler} for this datacenter.
-   * @param clusterMapConfig
-   */
-  CloudServiceClusterInfo(ClusterChangeHandler clusterChangeHandler, ClusterMapConfig clusterMapConfig) {
-    super(clusterChangeHandler);
-  }
-}
