+++ ambry-clustermap/src/main/java/com/github/ambry/clustermap/DatacenterInitializer.java
@@ -15,16 +15,12 @@
 
 package com.github.ambry.clustermap;
 
+import com.github.ambry.clustermap.HelixClusterManager.HelixClusterChangeHandler;
 import com.github.ambry.config.ClusterMapConfig;
 import com.github.ambry.utils.Utils;
-import java.nio.ByteBuffer;
-import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.atomic.AtomicLong;
 import org.apache.helix.HelixManager;
 import org.apache.helix.InstanceType;
 import org.apache.helix.PropertyType;
@@ -46,19 +42,10 @@ class DatacenterInitializer {
   private final HelixFactory helixFactory;
   private final ClusterMapUtils.DcZkInfo dcZkInfo;
   private final String dcName;
-
-  // Fields to pass into both ClusterChangeHandlers
+  private final HelixClusterManager helixClusterManager;
   private final String selfInstanceName;
-  private final Map<String, Map<String, String>> partitionOverrideInfoMap;
-  private final HelixClusterManager.ClusterChangeHandlerCallback clusterChangeHandlerCallback;
   private final HelixClusterManager.HelixClusterManagerCallback helixClusterManagerCallback;
-  private final HelixClusterManagerMetrics helixClusterManagerMetrics;
   private final DataNodeConfigSourceMetrics dataNodeConfigSourceMetrics;
-  private final AtomicLong sealedStateChangeCounter;
-  // Fields to pass into only SimpleClusterChangeHandler (These can be removed if SimpleClusterChangeHandler is removed)
-  private final ConcurrentHashMap<ByteBuffer, AmbryPartition> partitionMap;
-  private final ConcurrentHashMap<String, AmbryPartition> partitionNameToAmbryPartition;
-  private final ConcurrentHashMap<AmbryPartition, Set<AmbryReplica>> ambryPartitionToAmbryReplicas;
 
   /**
    * @param clusterMapConfig {@link ClusterMapConfig} to help some admin operations
@@ -66,40 +53,23 @@ class DatacenterInitializer {
    * @param helixFactory the {@link HelixFactory} instance to construct managers.
    * @param dcZkInfo info about the DC, like connection string, name, and replica type
    * @param selfInstanceName the name of instance on which {@link HelixClusterManager} resides.
-   * @param partitionOverrideInfoMap a map specifying partitions whose state should be overridden.
-   * @param clusterChangeHandlerCallback a call back that allows current handler to update cluster-wide info.
    * @param helixClusterManagerCallback a help class to get cluster state from all DCs.
-   * @param helixClusterManagerMetrics metrics that help track of cluster changes and infos.
    * @param dataNodeConfigSourceMetrics metrics related to {@link DataNodeConfigSource}.
-   * @param sealedStateChangeCounter a counter that records event when replica is sealed or unsealed
-   * @param partitionMap a map from serialized bytes to corresponding partition.
-   * @param partitionNameToAmbryPartition a map from partition name to {@link AmbryPartition} object.
-   * @param ambryPartitionToAmbryReplicas a map from {@link AmbryPartition} to its replicas.
+   * @param helixClusterManager helixClusterManager reference to {@link HelixClusterManager} that creates this class.
    */
   DatacenterInitializer(ClusterMapConfig clusterMapConfig, HelixManager localManager, HelixFactory helixFactory,
       ClusterMapUtils.DcZkInfo dcZkInfo, String selfInstanceName,
-      Map<String, Map<String, String>> partitionOverrideInfoMap,
-      HelixClusterManager.ClusterChangeHandlerCallback clusterChangeHandlerCallback,
       HelixClusterManager.HelixClusterManagerCallback helixClusterManagerCallback,
-      HelixClusterManagerMetrics helixClusterManagerMetrics, DataNodeConfigSourceMetrics dataNodeConfigSourceMetrics,
-      AtomicLong sealedStateChangeCounter, ConcurrentHashMap<ByteBuffer, AmbryPartition> partitionMap,
-      ConcurrentHashMap<String, AmbryPartition> partitionNameToAmbryPartition,
-      ConcurrentHashMap<AmbryPartition, Set<AmbryReplica>> ambryPartitionToAmbryReplicas) {
+      DataNodeConfigSourceMetrics dataNodeConfigSourceMetrics, HelixClusterManager helixClusterManager) {
     this.clusterMapConfig = clusterMapConfig;
     this.localManager = localManager;
     this.helixFactory = helixFactory;
     this.dcZkInfo = dcZkInfo;
     this.selfInstanceName = selfInstanceName;
-    this.partitionOverrideInfoMap = partitionOverrideInfoMap;
-    this.clusterChangeHandlerCallback = clusterChangeHandlerCallback;
     this.helixClusterManagerCallback = helixClusterManagerCallback;
-    this.helixClusterManagerMetrics = helixClusterManagerMetrics;
     this.dataNodeConfigSourceMetrics = dataNodeConfigSourceMetrics;
-    this.sealedStateChangeCounter = sealedStateChangeCounter;
-    this.partitionMap = partitionMap;
-    this.partitionNameToAmbryPartition = partitionNameToAmbryPartition;
-    this.ambryPartitionToAmbryReplicas = ambryPartitionToAmbryReplicas;
-    dcName = dcZkInfo.getDcName();
+    this.dcName = dcZkInfo.getDcName();
+    this.helixClusterManager = helixClusterManager;
   }
 
   /**
@@ -108,17 +78,10 @@ class DatacenterInitializer {
   public void start() {
     Utils.newThread(() -> {
       try {
-        DcInfo dcInfo;
-        switch (dcZkInfo.getReplicaType()) {
-          case DISK_BACKED:
-            dcInfo = initializeHelixDatacenter();
-            break;
-          case CLOUD_BACKED:
-            dcInfo = initializeCloudDatacenter();
-            break;
-          default:
-            throw new UnsupportedOperationException("Unknown replica type: " + dcZkInfo.getReplicaType());
+        if (dcZkInfo.getReplicaType() != ReplicaType.DISK_BACKED) {
+          throw new UnsupportedOperationException("Unknown replica type: " + dcZkInfo.getReplicaType());
         }
+        DcInfo dcInfo = initializeHelixDatacenter();
         initializationFuture.complete(dcInfo);
       } catch (Exception e) {
         logger.warn("Exception while initializing datacenter {}", dcName, e);
@@ -171,21 +134,11 @@ class DatacenterInitializer {
       manager = helixFactory.getZkHelixManagerAndConnect(clusterMapConfig.clusterMapClusterName, selfInstanceName,
           InstanceType.SPECTATOR, zkConnectStr);
     }
-    HelixClusterChangeHandler clusterChangeHandler;
-    String clusterChangeHandlerType = clusterMapConfig.clusterMapClusterChangeHandlerType;
-    if (clusterChangeHandlerType.equals(SimpleClusterChangeHandler.class.getSimpleName())) {
-      clusterChangeHandler =
-          new SimpleClusterChangeHandler(clusterMapConfig, dcName, selfInstanceName, partitionOverrideInfoMap,
-              partitionMap, partitionNameToAmbryPartition, ambryPartitionToAmbryReplicas, helixClusterManagerCallback,
-              helixClusterManagerMetrics, this::onInitializationFailure, sealedStateChangeCounter);
-    } else if (clusterChangeHandlerType.equals(DynamicClusterChangeHandler.class.getSimpleName())) {
-      clusterChangeHandler =
-          new DynamicClusterChangeHandler(clusterMapConfig, dcName, selfInstanceName, partitionOverrideInfoMap,
-              helixClusterManagerCallback, clusterChangeHandlerCallback, helixClusterManagerMetrics,
-              this::onInitializationFailure, sealedStateChangeCounter);
-    } else {
-      throw new IllegalArgumentException("Unsupported cluster change handler type: " + clusterChangeHandlerType);
-    }
+
+    HelixClusterChangeHandler clusterChangeHandler =
+        helixClusterManager.new HelixClusterChangeHandler(dcName, helixClusterManagerCallback,
+            this::onInitializationFailure, false);
+
     // Create RoutingTableProvider of each DC to keep track of partition(replicas) state. Here, we use current
     // state based RoutingTableProvider to remove dependency on Helix's pipeline and reduce notification latency.
     logger.info("Creating routing table provider associated with Helix manager at {}", zkConnectStr);
@@ -211,12 +164,12 @@ class DatacenterInitializer {
     logger.info("Registered live instance change listeners for Helix manager at {}", zkConnectStr);
 
     // in case initial event occurs before adding routing table listener, here we explicitly set snapshot in
-    // ClusterChangeHandler. The reason is, if listener missed initial event, snapshot inside routing table
+    // HelixClusterManager. The reason is, if listener missed initial event, snapshot inside routing table
     // provider should be already populated.
     clusterChangeHandler.setRoutingTableSnapshot(routingTableProvider.getRoutingTableSnapshot());
     // the initial routing table change should populate the instanceConfigs. If it's empty that means initial
     // change didn't come and thread should wait on the init latch to ensure routing table snapshot is non-empty
-    if (clusterChangeHandler.getRoutingTableSnapshot().getInstanceConfigs().isEmpty()) {
+    if (helixClusterManagerCallback.getRoutingTableSnapshot(dcName).getInstanceConfigs().isEmpty()) {
       // Periodic refresh in routing table provider is enabled by default. In worst case, routerUpdater should
       // trigger routing table change within 5 minutes
       logger.info("Routing table snapshot in {} is currently empty. Waiting for initial notification.", dcName);
@@ -229,19 +182,5 @@ class DatacenterInitializer {
     }
 
     return new HelixDcInfo(dcName, dcZkInfo, manager, clusterChangeHandler, dataNodeConfigSource);
-  }
-
-  /**
-   * Currently, this does not connect to the VCR zookeeper and assumes that all partitions are supported in the cloud
-   * datacenter. This will be the case until the VCR and native storage clusters are unified under the same
-   * {@link ClusterMap}. Once this happens, we can use the VCR cluster as a source of truth for supported partitions
-   * in the cloud datacenter.
-   * @return the {@link DcInfo} for the cloud datacenter.
-   * @throws Exception if something went wrong during startup
-   */
-  private DcInfo initializeCloudDatacenter() throws Exception {
-    CloudServiceClusterChangeHandler clusterChangeHandler =
-        new CloudServiceClusterChangeHandler(dcName, clusterMapConfig, clusterChangeHandlerCallback);
-    return new DcInfo(dcName, dcZkInfo, clusterChangeHandler);
   }
 }
