+++ ambry-clustermap/src/main/java/com/github/ambry/clustermap/HelixClusterManager.java
@@ -15,6 +15,7 @@ package com.github.ambry.clustermap;
 
 import com.codahale.metrics.MetricRegistry;
 import com.github.ambry.config.ClusterMapConfig;
+import com.github.ambry.utils.Pair;
 import com.github.ambry.utils.SystemTime;
 import java.io.IOException;
 import java.io.InputStream;
@@ -30,10 +31,21 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
 import org.apache.helix.AccessOption;
 import org.apache.helix.HelixManager;
 import org.apache.helix.InstanceType;
+import org.apache.helix.NotificationContext;
+import org.apache.helix.api.listeners.IdealStateChangeListener;
+import org.apache.helix.api.listeners.LiveInstanceChangeListener;
+import org.apache.helix.api.listeners.RoutingTableChangeListener;
+import org.apache.helix.model.IdealState;
+import org.apache.helix.model.LiveInstance;
 import org.apache.helix.spectator.RoutingTableSnapshot;
 import org.apache.helix.store.zk.ZkHelixPropertyStore;
 import org.apache.helix.zookeeper.datamodel.ZNRecord;
@@ -60,6 +72,15 @@ public class HelixClusterManager implements ClusterMap {
   private final ClusterMapConfig clusterMapConfig;
   private final Map<String, DcInfo> dcToDcInfo = new HashMap<>();
   private final Map<Byte, String> dcIdToDcName = new HashMap<>();
+  private final ConcurrentHashMap<String, Set<AmbryDataNode>> dcToNodes = new ConcurrentHashMap<>();
+  // A map whose key is ambry datanode and value is a map of partitionId to corresponding replica associated with this datanode
+  private final ConcurrentHashMap<AmbryDataNode, ConcurrentHashMap<String, AmbryReplica>> ambryDataNodeToAmbryReplicas =
+      new ConcurrentHashMap<>();
+  private final ConcurrentHashMap<AmbryDataNode, Set<AmbryDisk>> ambryDataNodeToAmbryDisks = new ConcurrentHashMap<>();
+  private final Map<String, ConcurrentHashMap<String, String>> partitionToResourceNameByDc = new ConcurrentHashMap<>();
+  private final Map<String, AtomicReference<RoutingTableSnapshot>> dcToRoutingTableSnapshotRef = new HashMap<>();
+  private final AtomicReference<RoutingTableSnapshot> globalRoutingTableSnapshotRef = new AtomicReference<>();
+  private final ConcurrentHashMap<String, AmbryDataNode> instanceNameToAmbryDataNode = new ConcurrentHashMap<>();
   private final ConcurrentHashMap<String, AmbryPartition> partitionNameToAmbryPartition = new ConcurrentHashMap<>();
   private final ConcurrentHashMap<AmbryPartition, Set<AmbryReplica>> ambryPartitionToAmbryReplicas =
       new ConcurrentHashMap<>();
@@ -68,7 +89,6 @@ public class HelixClusterManager implements ClusterMap {
   private final AtomicLong clusterWideAllocatedRawCapacityBytes = new AtomicLong(0);
   private final AtomicLong clusterWideAllocatedUsableCapacityBytes = new AtomicLong(0);
   private final HelixClusterManagerCallback helixClusterManagerCallback;
-  private final ClusterChangeHandlerCallback clusterChangeHandlerCallback;
   private final byte localDatacenterId;
   private final AtomicLong sealedStateChangeCounter = new AtomicLong(0);
   private final PartitionSelectionHelper partitionSelectionHelper;
@@ -80,9 +100,9 @@ public class HelixClusterManager implements ClusterMap {
   // reflects the current value.
   private final AtomicLong currentXid;
   final HelixClusterManagerMetrics helixClusterManagerMetrics;
-  private ClusterInfo clusterInfo = null;
-  private final boolean clusterMapUseAggregatedView = false;
   Map<String, DcZkInfo> dataCenterToZkAddress = null;
+  private ClusterInfo clusterInfo = null;
+  private final AtomicLong errorCount = new AtomicLong(0);
 
   /**
    * Instantiate a HelixClusterManager.
@@ -101,7 +121,6 @@ public class HelixClusterManager implements ClusterMap {
     clusterName = clusterMapConfig.clusterMapClusterName;
     selfInstanceName = instanceName;
     helixClusterManagerCallback = new HelixClusterManagerCallback();
-    clusterChangeHandlerCallback = new ClusterChangeHandlerCallback();
     helixClusterManagerMetrics = new HelixClusterManagerMetrics(metricRegistry, helixClusterManagerCallback);
     HelixManager localManager = null;
     Map<String, Exception> initializationFailureMap = new HashMap<>();
@@ -119,11 +138,14 @@ public class HelixClusterManager implements ClusterMap {
         try {
           ClusterInitializer clusterInitializer =
               new ClusterInitializer(clusterMapConfig, dataCenterToZkAddress, helixFactory, selfInstanceName,
-                  partitionOverrideInfoMap, new ClusterChangeHandlerCallback(), helixClusterManagerCallback,
-                  helixClusterManagerMetrics, dataNodeConfigSourceMetrics, sealedStateChangeCounter, partitionMap,
-                  partitionNameToAmbryPartition, ambryPartitionToAmbryReplicas);
+                  helixClusterManagerCallback, dataNodeConfigSourceMetrics, this);
           clusterInitializer.start();
           clusterInfo = clusterInitializer.join();
+          // Populate dcToDCInfo and dcIdToDcName maps for all DCs if cluster initialization succeeds.
+          for (DcZkInfo dcZkInfo : dataCenterToZkAddress.values()) {
+            dcToDcInfo.put(dcZkInfo.getDcName(), new DcInfo(dcZkInfo.getDcName(), dcZkInfo, null));
+            dcIdToDcName.put(dcZkInfo.getDcId(), dcZkInfo.getDcName());
+          }
         } catch (Exception e) {
           initializationFailureMap.putIfAbsent(clusterMapConfig.clusterMapDatacenterName, e);
         }
@@ -133,9 +155,7 @@ public class HelixClusterManager implements ClusterMap {
           // Initialize from every remote datacenter in a separate thread to speed things up.
           DatacenterInitializer initializer =
               new DatacenterInitializer(clusterMapConfig, localManager, helixFactory, dcZkInfo, selfInstanceName,
-                  partitionOverrideInfoMap, new ClusterChangeHandlerCallback(), helixClusterManagerCallback,
-                  helixClusterManagerMetrics, dataNodeConfigSourceMetrics, sealedStateChangeCounter, partitionMap,
-                  partitionNameToAmbryPartition, ambryPartitionToAmbryReplicas);
+                  helixClusterManagerCallback, dataNodeConfigSourceMetrics, this);
           initializer.start();
           initializers.add(initializer);
         }
@@ -146,20 +166,6 @@ public class HelixClusterManager implements ClusterMap {
             dcIdToDcName.put(dcInfo.dcZkInfo.getDcId(), dcInfo.dcName);
           } catch (Exception e) {
             initializationFailureMap.putIfAbsent(initializer.getDcName(), e);
-          }
-        }
-
-        for (DcInfo dcInfo : dcToDcInfo.values()) {
-          if (dcInfo.clusterChangeHandler instanceof ClusterMapChangeListener) {
-            ClusterMapChangeListener listener = (ClusterMapChangeListener) dcInfo.clusterChangeHandler;
-            registerClusterMapListener(listener);
-            // WARNING: currently this code is tailored to the CloudServiceClusterChangeHandler, which only needs to be
-            // provided one replica per partition. If this assumption is no longer valid, modify this logic.
-            List<ReplicaId> oneReplicaPerPartition = new ArrayList<>(ambryPartitionToAmbryReplicas.size());
-            for (Set<AmbryReplica> replicas : ambryPartitionToAmbryReplicas.values()) {
-              replicas.stream().findFirst().ifPresent(oneReplicaPerPartition::add);
-            }
-            listener.onReplicaAddedOrRemoved(oneReplicaPerPartition, Collections.emptyList());
           }
         }
       }
@@ -176,11 +182,6 @@ public class HelixClusterManager implements ClusterMap {
       // resolve the status of all partitions before completing initialization
       for (AmbryPartition partition : partitionMap.values()) {
         partition.resolvePartitionState();
-      }
-      if (clusterMapConfig.clusterMapClusterChangeHandlerType.equals(
-          SimpleClusterChangeHandler.class.getSimpleName())) {
-        // capacity stats needs to be initialized only when SimpleClusterChangeHandler is adopted.
-        initializeCapacityStats();
       }
       helixClusterManagerMetrics.initializeInstantiationMetric(true,
           initializationFailureMap.values().stream().filter(Objects::nonNull).count());
@@ -252,41 +253,8 @@ public class HelixClusterManager implements ClusterMap {
     return manager;
   }
 
-  /**
-   * Initialize capacity statistics.
-   */
-  private void initializeCapacityStats() {
-    if (clusterMapUseAggregatedView) {
-      Map<AmbryDataNode, Set<AmbryDisk>> dataNodeToDisks = clusterInfo.clusterChangeHandler.getDataNodeToDisksMap();
-      for (Set<AmbryDisk> disks : dataNodeToDisks.values()) {
-        for (AmbryDisk disk : disks) {
-          clusterWideRawCapacityBytes.getAndAdd(disk.getRawCapacityInBytes());
-        }
-      }
-    } else {
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        Map<AmbryDataNode, Set<AmbryDisk>> dataNodeToDisks = dcInfo.clusterChangeHandler.getDataNodeToDisksMap();
-        for (Set<AmbryDisk> disks : dataNodeToDisks.values()) {
-          for (AmbryDisk disk : disks) {
-            clusterWideRawCapacityBytes.getAndAdd(disk.getRawCapacityInBytes());
-          }
-        }
-      }
-    }
-    for (Set<AmbryReplica> partitionReplicas : ambryPartitionToAmbryReplicas.values()) {
-      long replicaCapacity = partitionReplicas.iterator().next().getCapacityInBytes();
-      clusterWideAllocatedRawCapacityBytes.getAndAdd(replicaCapacity * partitionReplicas.size());
-      clusterWideAllocatedUsableCapacityBytes.getAndAdd(replicaCapacity);
-    }
-  }
-
   @Override
   public boolean hasDatacenter(String datacenterName) {
-    if (clusterMapUseAggregatedView) {
-      //TODO: This is a static check since it could happen that a data center is not reachable and is missing in
-      // aggregated view information coming from helix.
-      return dataCenterToZkAddress.containsKey(datacenterName);
-    }
     return dcToDcInfo.containsKey(datacenterName);
   }
 
@@ -302,21 +270,7 @@ public class HelixClusterManager implements ClusterMap {
 
   @Override
   public AmbryDataNode getDataNodeId(String hostname, int port) {
-    String instanceName = getInstanceName(hostname, port);
-
-    if (clusterMapUseAggregatedView) {
-      return clusterInfo.clusterChangeHandler.getDataNode(instanceName);
-    }
-
-    AmbryDataNode dataNode = null;
-    for (DcInfo dcInfo : dcToDcInfo.values()) {
-      dataNode = dcInfo.clusterChangeHandler.getDataNode(instanceName);
-      if (dataNode != null) {
-        break;
-      }
-    }
-
-    return dataNode;
+    return instanceNameToAmbryDataNode.get(getInstanceName(hostname, port));
   }
 
   @Override
@@ -324,25 +278,12 @@ public class HelixClusterManager implements ClusterMap {
     if (!(dataNodeId instanceof AmbryDataNode)) {
       throw new IllegalArgumentException("Incompatible type passed in");
     }
-    AmbryDataNode dataNode = (AmbryDataNode) dataNodeId;
-    if (clusterMapUseAggregatedView) {
-      return clusterInfo.clusterChangeHandler.getReplicaIds(dataNode);
-    } else {
-      return dcToDcInfo.get(dataNode.getDatacenterName()).clusterChangeHandler.getReplicaIds(dataNode);
-    }
+    return new ArrayList<>(ambryDataNodeToAmbryReplicas.get(dataNodeId).values());
   }
 
   @Override
   public List<AmbryDataNode> getDataNodeIds() {
-    List<AmbryDataNode> dataNodeList = new ArrayList<>();
-    if (clusterMapUseAggregatedView) {
-      dataNodeList.addAll(clusterInfo.clusterChangeHandler.getAllDataNodes());
-    } else {
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        dataNodeList.addAll(dcInfo.clusterChangeHandler.getAllDataNodes());
-      }
-    }
-    return dataNodeList;
+    return new ArrayList<>(instanceNameToAmbryDataNode.values());
   }
 
   @Override
@@ -391,13 +332,12 @@ public class HelixClusterManager implements ClusterMap {
     snapshot.put(CLUSTER_NAME, clusterName);
     snapshot.put(TIMESTAMP_MS, SystemTime.getInstance().milliseconds());
     JSONArray datacentersJsonArray = new JSONArray();
-    dcIdToDcName.forEach((dcId, dcName) -> {
+    dataCenterToZkAddress.forEach((dc, zkInfo) -> {
       JSONObject data = new JSONObject();
-      data.put(DATACENTER_NAME, dcName);
-      data.put(DATACENTER_ID, dcId);
+      data.put(DATACENTER_NAME, zkInfo.getDcName());
+      data.put(DATACENTER_ID, zkInfo.getDcId());
       JSONArray datanodesInDc = new JSONArray();
-      dcToDcInfo.get(dcName).clusterChangeHandler.getAllDataNodes()
-          .forEach(node -> datanodesInDc.put(node.getSnapshot()));
+      dcToNodes.get(zkInfo.getDcName()).forEach(node -> datanodesInDc.put(node.getSnapshot()));
       data.put(DATACENTER_NODES, datanodesInDc);
       datacentersJsonArray.put(data);
     });
@@ -490,19 +430,12 @@ public class HelixClusterManager implements ClusterMap {
       currentPartition = mappedPartition;
     }
     // Check if data node or disk is in current cluster map, if not, set bootstrapReplica to null.
-    ClusterChangeHandler clusterChangeHandler;
-    if (clusterMapUseAggregatedView) {
-      clusterChangeHandler = clusterInfo.clusterChangeHandler;
-    } else {
-      clusterChangeHandler = dcToDcInfo.get(clusterMapConfig.clusterMapDatacenterName).clusterChangeHandler;
-    }
-
-    AmbryDataNode dataNode = clusterChangeHandler.getDataNode(instanceName);
+    AmbryDataNode dataNode = instanceNameToAmbryDataNode.get(instanceName);
     String mountPathAndDiskCapacityFromHelix = replicaInfos.get(instanceName);
     String[] segments = mountPathAndDiskCapacityFromHelix.split(DISK_CAPACITY_DELIM_STR);
     String mountPath = segments[0];
     String diskCapacityStr = segments.length >= 2 ? segments[1] : null;
-    Set<AmbryDisk> disks = dataNode != null ? clusterChangeHandler.getDisks(dataNode) : null;
+    Set<AmbryDisk> disks = dataNode != null ? ambryDataNodeToAmbryDisks.get(dataNode) : null;
     Optional<AmbryDisk> potentialDisk =
         disks != null ? disks.stream().filter(d -> d.getMountPath().equals(mountPath)).findAny() : Optional.empty();
     if (potentialDisk.isPresent()) {
@@ -537,7 +470,7 @@ public class HelixClusterManager implements ClusterMap {
 
   @Override
   public void registerClusterMapListener(ClusterMapChangeListener clusterMapChangeListener) {
-    if (clusterMapUseAggregatedView) {
+    if (clusterMapConfig.clusterMapUseAggregatedView) {
       clusterInfo.clusterChangeHandler.registerClusterMapListener(clusterMapChangeListener);
     } else {
       for (DcInfo dcInfo : dcToDcInfo.values()) {
@@ -551,7 +484,7 @@ public class HelixClusterManager implements ClusterMap {
    */
   @Override
   public void close() {
-    if (clusterMapUseAggregatedView) {
+    if (clusterMapConfig.clusterMapUseAggregatedView) {
       clusterInfo.close();
     } else {
       for (DcInfo dcInfo : dcToDcInfo.values()) {
@@ -570,27 +503,14 @@ public class HelixClusterManager implements ClusterMap {
   AmbryReplica getReplicaForPartitionOnNode(DataNodeId dataNodeId, String partitionString) {
     // Note: partitionString here is now from partitionId.toPathString()
     AmbryDataNode ambryDataNode = getDataNodeId(dataNodeId.getHostname(), dataNodeId.getPort());
-    if (clusterMapUseAggregatedView) {
-      return clusterInfo.clusterChangeHandler.getReplicaId(ambryDataNode, partitionString);
-    } else {
-      return dcToDcInfo.get(dataNodeId.getDatacenterName()).clusterChangeHandler.getReplicaId(ambryDataNode,
-          partitionString);
-    }
+    return ambryDataNodeToAmbryReplicas.getOrDefault(ambryDataNode, new ConcurrentHashMap<>()).get(partitionString);
   }
 
   /**
    * @return the count of errors encountered by the Cluster Manager.
    */
   long getErrorCount() {
-    long totalErrorCnt = 0;
-    if (clusterMapUseAggregatedView) {
-      totalErrorCnt = clusterInfo.clusterChangeHandler.getErrorCount();
-    } else {
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        totalErrorCnt += dcInfo.clusterChangeHandler.getErrorCount();
-      }
-    }
-    return totalErrorCnt;
+    return errorCount.get();
   }
 
   /**
@@ -605,17 +525,7 @@ public class HelixClusterManager implements ClusterMap {
    * @return a map of partition to its corresponding resource grouped by data center
    */
   Map<String, Map<String, String>> getPartitionToResourceMap() {
-    Map<String, Map<String, String>> partitionToResourceNameByDc = new HashMap<>();
-    if (clusterMapUseAggregatedView) {
-      //TODO: check if all data centers needs to be populated
-      partitionToResourceNameByDc.put(clusterMapConfig.clusterMapDatacenterName,
-          clusterInfo.clusterChangeHandler.getPartitionToResourceMap());
-    } else {
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        partitionToResourceNameByDc.put(dcInfo.dcName, dcInfo.clusterChangeHandler.getPartitionToResourceMap());
-      }
-    }
-    return partitionToResourceNameByDc;
+    return Collections.unmodifiableMap(partitionToResourceNameByDc);
   }
 
   /**
@@ -623,16 +533,6 @@ public class HelixClusterManager implements ClusterMap {
    * @return a map of data center to its data nodes
    */
   Map<String, Set<AmbryDataNode>> getDcToDataNodesMap() {
-    Map<String, Set<AmbryDataNode>> dcToNodes = new HashMap<>();
-    if (clusterMapUseAggregatedView) {
-      clusterInfo.clusterChangeHandler.getAllDataNodes()
-          .forEach(ambryDataNode -> dcToNodes.computeIfAbsent(ambryDataNode.getDatacenterName(), k -> new HashSet<>())
-              .add(ambryDataNode));
-    } else {
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        dcToNodes.put(dcInfo.dcName, new HashSet<>(dcInfo.clusterChangeHandler.getAllDataNodes()));
-      }
-    }
     return Collections.unmodifiableMap(dcToNodes);
   }
 
@@ -642,14 +542,12 @@ public class HelixClusterManager implements ClusterMap {
    */
   Map<String, RoutingTableSnapshot> getRoutingTableSnapshots() {
     Map<String, RoutingTableSnapshot> dcToRoutingTableSnapshot = new HashMap<>();
-    if (clusterMapUseAggregatedView) {
-      // TODO
+    if (clusterMapConfig.clusterMapUseAggregatedView) {
+      // TODO Fix tests
+      dcToRoutingTableSnapshot.put(clusterMapConfig.clusterMapDatacenterName, globalRoutingTableSnapshotRef.get());
     } else {
       for (DcInfo dcInfo : dcToDcInfo.values()) {
-        ClusterChangeHandler handler = dcInfo.clusterChangeHandler;
-        if (handler instanceof HelixClusterChangeHandler) {
-          dcToRoutingTableSnapshot.put(dcInfo.dcName, ((HelixClusterChangeHandler) handler).getRoutingTableSnapshot());
-        }
+        dcToRoutingTableSnapshot.put(dcInfo.dcName, dcToRoutingTableSnapshotRef.get(dcInfo.dcName).get());
       }
     }
     return Collections.unmodifiableMap(dcToRoutingTableSnapshot);
@@ -674,90 +572,554 @@ public class HelixClusterManager implements ClusterMap {
    * @return localDC's zk connect string.
    */
   public String getLocalDcZkConnectString() {
-    // TODO
     return dcToDcInfo.get(clusterMapConfig.clusterMapDatacenterName).dcZkInfo.getZkConnectStrs().get(0);
   }
 
   /**
-   * A callback class for {@link HelixClusterChangeHandler} in each dc to update cluster-wide info (i.e partition-to-replica
-   * mapping, cluster-wide capacity)
+   * Add partition if it's not present in cluster-wide partition map and also update cluster-wide allocated usable
+   * capacity. If the partition already exists, skip addition and return current partition.
+   * @param partition the {@link AmbryPartition} to add (if not present)
+   * @param capacityBytes the capacity of partition in bytes
+   * @return the current {@link AmbryPartition} present in the map.
    */
-  class ClusterChangeHandlerCallback {
+  private AmbryPartition addPartitionIfAbsent(AmbryPartition partition, long capacityBytes) {
+    AmbryPartition currentPartition = partitionNameToAmbryPartition.putIfAbsent(partition.toPathString(), partition);
+    if (currentPartition == null) {
+      // this means the map previously didn't contain this partition and passed-in partition is successfully added
+      // into the map
+      currentPartition = partition;
+      // it doesn't really need to synchronize this method. "partitionNameToAmbryPartition" guarantees each thread
+      // will get same instance of ambry partition. The first one that succeeds adding partition into
+      // "partitionNameToAmbryPartition" will update "partitionMap".
+      partitionMap.put(ByteBuffer.wrap(currentPartition.getBytes()), currentPartition);
+      // update cluster-wide capacity
+      clusterWideAllocatedUsableCapacityBytes.getAndAdd(capacityBytes);
+    }
+    ambryPartitionToAmbryReplicas.putIfAbsent(currentPartition, ConcurrentHashMap.newKeySet());
+    return currentPartition;
+  }
+
+  /**
+   * Add a list of {@link AmbryReplica} to given {@link AmbryPartition} and update cluster-wide capacity stats
+   * @param partition the {@link AmbryPartition} which replicas should be added to.
+   * @param replicas list of {@link AmbryReplica} to be added.
+   */
+  private void addReplicasToPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
+    AmbryPartition currentPartition = addPartitionIfAbsent(partition, replicas.get(0).getCapacityInBytes());
+    ambryPartitionToAmbryReplicas.compute(currentPartition, (k, v) -> {
+      // calling addPartitionIfAbsent guarantees that v is not null
+      v.addAll(replicas);
+      return v;
+    });
+    clusterWideAllocatedRawCapacityBytes.getAndAdd(replicas.get(0).getCapacityInBytes() * replicas.size());
+  }
+
+  /**
+   * Remove {@link AmbryReplica}(s) from given {@link AmbryPartition} and update cluster-wide capacity stats
+   * @param partition the {@link AmbryPartition} which replicas should be removed from.
+   * @param replicas list of {@link AmbryReplica} to be removed.
+   */
+  private void removeReplicasFromPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
+    ambryPartitionToAmbryReplicas.computeIfPresent(partition, (k, v) -> {
+      v.removeAll(replicas);
+      clusterWideAllocatedRawCapacityBytes.getAndAdd(-1 * replicas.get(0).getCapacityInBytes() * replicas.size());
+      return v;
+    });
+  }
+
+  /**
+   * Add cluster-wide raw capacity. This is called when new disk is created and added to cluster. We update
+   * cluster-wide raw capacity by adding the disk capacity.
+   * @param diskRawCapacityBytes raw disk capacity to be added.
+   */
+  private void addClusterWideRawCapacity(long diskRawCapacityBytes) {
+    clusterWideRawCapacityBytes.getAndAdd(diskRawCapacityBytes);
+  }
+
+  /**
+   * Pop out bootstrap replica (if any) on current instance. A bootstrap replica is a replica dynamically added to
+   * current node at runtime.
+   * @param partitionName the partition name of bootstrap replica.
+   * @return bootstrap replica or {@code null} if not found.
+   */
+  private AmbryReplica fetchBootstrapReplica(String partitionName) {
+    return (AmbryReplica) bootstrapReplicas.remove(partitionName);
+  }
+
+  /**
+   * An instance of this object is used to register as listener for Helix related changes in each datacenter. This
+   * class is also responsible for handling events received.
+   */
+  class HelixClusterChangeHandler
+      implements DataNodeConfigChangeListener, LiveInstanceChangeListener, IdealStateChangeListener,
+                 RoutingTableChangeListener {
+    private final String dcOrClusterName;
+    final Set<String> allInstances = new HashSet<>();
+    private final Object notificationLock = new Object();
+    private final Consumer<Exception> onInitializationFailure;
+    private final boolean isAggregatedViewHandler;
+    private final HelixClusterManager.HelixClusterManagerCallback helixClusterManagerCallback;
+    private final CountDownLatch routingTableInitLatch = new CountDownLatch(1);
+    // A map whose key is ambry datanode and value is a map of partitionName to corresponding replica on this datanode
+    // Note: the partitionName (inner map key) comes from partitionId.toString() not partitionId.toPathString()
+    private final List<ClusterMapChangeListener> clusterMapChangeListeners = new ArrayList<>();
+
+    private volatile boolean instanceConfigInitialized = false;
+    private volatile boolean liveStateInitialized = false;
+    private volatile boolean idealStateInitialized = false;
 
     /**
-     * Add partition if it's not present in cluster-wide partition map and also update cluster-wide allocated usable
-     * capacity. If the partition already exists, skip addition and return current partition.
-     * @param partition the {@link AmbryPartition} to add (if not present)
-     * @param capacityBytes the capacity of partition in bytes
-     * @return the current {@link AmbryPartition} present in the map.
+     * Instantiate a HelixClusterChangeHandler
+     * @param dcOrClusterName the name of data center this handler is associated with.
+     * @param helixClusterManagerCallback a call back used to query cluster-wide info.
+     * @param onInitializationFailure callback to be called if initialization fails in a listener call.
+     * @param isAggregatedViewHandler indicates if this is a handler for aggregated cluster
      */
-    AmbryPartition addPartitionIfAbsent(AmbryPartition partition, long capacityBytes) {
-      AmbryPartition currentPartition = partitionNameToAmbryPartition.putIfAbsent(partition.toPathString(), partition);
-      if (currentPartition == null) {
-        // this means the map previously didn't contain this partition and passed-in partition is successfully added
-        // into the map
-        currentPartition = partition;
-        // it doesn't really need to synchronize this method. "partitionNameToAmbryPartition" guarantees each thread
-        // will get same instance of ambry partition. The first one that succeeds adding partition into
-        // "partitionNameToAmbryPartition" will update "partitionMap".
-        partitionMap.put(ByteBuffer.wrap(currentPartition.getBytes()), currentPartition);
-        // update cluster-wide capacity
-        clusterWideAllocatedUsableCapacityBytes.getAndAdd(capacityBytes);
+    HelixClusterChangeHandler(String dcOrClusterName, HelixClusterManagerCallback helixClusterManagerCallback,
+        Consumer<Exception> onInitializationFailure, boolean isAggregatedViewHandler) {
+      this.dcOrClusterName = dcOrClusterName;
+      this.helixClusterManagerCallback = helixClusterManagerCallback;
+      this.onInitializationFailure = onInitializationFailure;
+      this.isAggregatedViewHandler = isAggregatedViewHandler;
+    }
+
+    /**
+     * Handle any {@link DataNodeConfig} related change in current datacenter. Several events will trigger instance config
+     * change: (1) replica's seal or stop state has changed; (2) new node or new partition is added; (3) new replica is
+     * added to existing node; (4) old replica is removed from existing node; (5) data node is deleted from cluster.
+     * For now, this handler supports (1)~(4). We may consider supporting (5) in the future.
+     * (The ZNode path of instance config in Helix is [AmbryClusterName]/CONFIGS/PARTICIPANT/[hostname_port])
+     * @param configs all the {@link DataNodeConfig}(s) in current data center. (Note that PreFetch is enabled by default
+     *                in Helix, which means all instance configs under "participants" ZNode will be sent to this method)
+     */
+    @Override
+    public void onDataNodeConfigChange(Iterable<DataNodeConfig> configs) {
+      try {
+        synchronized (notificationLock) {
+          if (!instanceConfigInitialized) {
+            logger.info("Received initial notification for instance config change from {}", dcOrClusterName);
+          } else {
+            logger.info("Instance config change triggered from {}", dcOrClusterName);
+          }
+          if (logger.isDebugEnabled()) {
+            configs.forEach(config -> logger.debug("Detailed data node config in {} is: {}", dcOrClusterName, config));
+          }
+          try {
+            addOrUpdateInstanceInfos(configs);
+          } catch (Exception e) {
+            if (!instanceConfigInitialized) {
+              logger.error("Exception occurred when initializing instances in {}: ", dcOrClusterName, e);
+              onInitializationFailure.accept(e);
+            } else {
+              logger.error("Exception occurred at runtime when handling instance config changes in {}: ",
+                  dcOrClusterName, e);
+              helixClusterManagerMetrics.instanceConfigChangeErrorCount.inc();
+            }
+          } finally {
+            instanceConfigInitialized = true;
+          }
+          sealedStateChangeCounter.incrementAndGet();
+          helixClusterManagerMetrics.instanceConfigChangeTriggerCount.inc();
+        }
+      } catch (Throwable t) {
+        errorCount.incrementAndGet();
+        throw t;
       }
-      ambryPartitionToAmbryReplicas.putIfAbsent(currentPartition, ConcurrentHashMap.newKeySet());
-      return currentPartition;
     }
 
     /**
-     * Add a list of {@link AmbryReplica} to given {@link AmbryPartition} and update cluster-wide capacity stats
-     * @param partition the {@link AmbryPartition} which replicas should be added to.
-     * @param replicas list of {@link AmbryReplica} to be added.
+     * Triggered whenever the IdealState in current data center has changed (for now, it is usually updated by Helix
+     * Bootstrap tool).
+     * @param idealState a list of {@link IdealState} that specifies ideal location of replicas.
+     * @param changeContext the {@link NotificationContext} associated.
      */
-    void addReplicasToPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
-      AmbryPartition currentPartition = addPartitionIfAbsent(partition, replicas.get(0).getCapacityInBytes());
-      ambryPartitionToAmbryReplicas.compute(currentPartition, (k, v) -> {
-        // calling addPartitionIfAbsent guarantees that v is not null
-        v.addAll(replicas);
-        return v;
+    @Override
+    public void onIdealStateChange(List<IdealState> idealState, NotificationContext changeContext) {
+      if (!idealStateInitialized) {
+        logger.info("Received initial notification for IdealState change from {}", dcOrClusterName);
+        idealStateInitialized = true;
+      } else {
+        logger.info("IdealState change triggered from {}", dcOrClusterName);
+      }
+      logger.debug("Detailed ideal states in {} are: {}", dcOrClusterName, idealState);
+      // rebuild the entire partition-to-resource map in current dc
+      ConcurrentHashMap<String, String> partitionToResourceMap = new ConcurrentHashMap<>();
+      for (IdealState state : idealState) {
+        String resourceName = state.getResourceName();
+        state.getPartitionSet().forEach(partitionName -> partitionToResourceMap.put(partitionName, resourceName));
+      }
+      partitionToResourceNameByDc.put(dcOrClusterName, partitionToResourceMap);
+      helixClusterManagerMetrics.idealStateChangeTriggerCount.inc();
+    }
+
+    /**
+     * Triggered whenever there is a change in the list of live instances.
+     * @param liveInstances the list of all live instances (not a change set) at the time of this call.
+     * @param changeContext the {@link NotificationContext} associated.
+     */
+    @Override
+    public void onLiveInstanceChange(List<LiveInstance> liveInstances, NotificationContext changeContext) {
+      try {
+        if (!liveStateInitialized) {
+          logger.info("Received initial notification for live instance change from {}", dcOrClusterName);
+          liveStateInitialized = true;
+        } else {
+          logger.info("Live instance change triggered from {}", dcOrClusterName);
+        }
+        logger.debug("Detailed live instances in {} are: {}", dcOrClusterName, liveInstances);
+        synchronized (notificationLock) {
+          updateInstanceLiveness(liveInstances);
+          helixClusterManagerMetrics.liveInstanceChangeTriggerCount.inc();
+        }
+      } catch (Throwable t) {
+        errorCount.incrementAndGet();
+        throw t;
+      }
+    }
+
+    /**
+     * Triggered whenever the state of replica in cluster has changed. The snapshot contains up-to-date state of all
+     * resources(replicas) in this data center.
+     * @param routingTableSnapshot a snapshot of routing table for this data center.
+     * @param context additional context associated with this change.
+     */
+    @Override
+    public void onRoutingTableChange(RoutingTableSnapshot routingTableSnapshot, Object context) {
+      if (isAggregatedViewHandler) {
+        globalRoutingTableSnapshotRef.getAndSet(routingTableSnapshot);
+      } else {
+        dcToRoutingTableSnapshotRef.computeIfAbsent(dcOrClusterName, k -> new AtomicReference<>())
+            .getAndSet(routingTableSnapshot);
+      }
+      if (routingTableInitLatch.getCount() == 1) {
+        logger.info("Received initial notification for routing table change from {}", dcOrClusterName);
+        routingTableInitLatch.countDown();
+      } else {
+        logger.info("Routing table change triggered from {}", dcOrClusterName);
+      }
+      //we should notify routing table change indication to different cluster map change listeners
+      //(i.e replication manager, partition selection helper, etc). However, only replication manager handles this
+      //callback now. On receiving this notification, listeners can query the latest state information of replicas by
+      // querying APIs provided in class RoutingTableSnapshot
+      for (ClusterMapChangeListener listener : clusterMapChangeListeners) {
+        listener.onRoutingTableChange();
+      }
+
+      helixClusterManagerMetrics.routingTableChangeTriggerCount.inc();
+    }
+
+    /**
+     * Wait for initial notification during startup.
+     */
+    public void waitForInitNotification() throws InterruptedException {
+      // wait slightly more than 5 mins to ensure routerUpdater refreshes the snapshot.
+      if (!routingTableInitLatch.await(320, TimeUnit.SECONDS)) {
+        throw new IllegalStateException(
+            "Initial routing table change from " + dcOrClusterName + " didn't come within 5 mins");
+      }
+    }
+
+    /**
+     * Register a listener of cluster map for any changes.
+     * @param clusterMapChangeListener the {@link ClusterMapChangeListener} to add.
+     */
+    public void registerClusterMapListener(ClusterMapChangeListener clusterMapChangeListener) {
+      clusterMapChangeListeners.add(clusterMapChangeListener);
+    }
+
+    /**
+     * Sets the helix {@link RoutingTableSnapshot} for entire cluster or a given data center.
+     * @param routingTableSnapshot snapshot of cluster mappings.
+     */
+    public void setRoutingTableSnapshot(RoutingTableSnapshot routingTableSnapshot) {
+      if (isAggregatedViewHandler) {
+        globalRoutingTableSnapshotRef.getAndSet(routingTableSnapshot);
+      } else {
+        dcToRoutingTableSnapshotRef.computeIfAbsent(dcOrClusterName, k -> new AtomicReference<>())
+            .getAndSet(routingTableSnapshot);
+      }
+    }
+
+    /**
+     * Add new instances or update existing instances based on {@link DataNodeConfig}(s). This may also invoke callbacks
+     * in some clustermap change listeners (i.e. {@link PartitionSelectionHelper}, ReplicationManager)
+     * @param dataNodeConfigs the {@link DataNodeConfig}(s) used to update in-mem cluster map.
+     */
+    private void addOrUpdateInstanceInfos(Iterable<DataNodeConfig> dataNodeConfigs) throws Exception {
+      List<ReplicaId> totalAddedReplicas = new ArrayList<>();
+      List<ReplicaId> totalRemovedReplicas = new ArrayList<>();
+      for (DataNodeConfig dataNodeConfig : dataNodeConfigs) {
+        Pair<List<ReplicaId>, List<ReplicaId>> addedAndRemovedReplicas;
+        if (instanceNameToAmbryDataNode.containsKey(dataNodeConfig.getInstanceName())) {
+          addedAndRemovedReplicas = updateInstanceInfo(dataNodeConfig);
+        } else {
+          addedAndRemovedReplicas = new Pair<>(createNewInstance(dataNodeConfig), new ArrayList<>());
+        }
+        totalAddedReplicas.addAll(addedAndRemovedReplicas.getFirst());
+        totalRemovedReplicas.addAll(addedAndRemovedReplicas.getSecond());
+      }
+      // if this is not initial InstanceConfig change and any replicas are added or removed, we should invoke callbacks
+      // for different clustermap change listeners (i.e replication manager, partition selection helper)
+      logger.info(
+          "In total, {} replicas are being added and {} replicas are being removed. instanceConfigInitialized: {}",
+          totalAddedReplicas.size(), totalRemovedReplicas.size(), instanceConfigInitialized);
+      if (instanceConfigInitialized && (!totalAddedReplicas.isEmpty() || !totalRemovedReplicas.isEmpty())) {
+        for (ClusterMapChangeListener listener : clusterMapChangeListeners) {
+          listener.onReplicaAddedOrRemoved(totalAddedReplicas, totalRemovedReplicas);
+        }
+      }
+    }
+
+    /**
+     * Update info of an existing instance. This may happen in following cases: (1) new replica is added; (2) old replica
+     * is removed; (3) replica's state has changed (i.e. becomes seal/unseal).
+     * @param dataNodeConfig the {@link DataNodeConfig} used to update info of instance.
+     * @return a pair of lists: (1) new added replicas; (2) removed old replicas, during this update.
+     */
+    private Pair<List<ReplicaId>, List<ReplicaId>> updateInstanceInfo(DataNodeConfig dataNodeConfig) throws Exception {
+      final List<ReplicaId> addedReplicas = new ArrayList<>();
+      final List<ReplicaId> removedReplicas = new ArrayList<>();
+      String instanceName = dataNodeConfig.getInstanceName();
+      logger.info("Updating replicas info for existing node {}", instanceName);
+      Set<String> sealedReplicas = dataNodeConfig.getSealedReplicas();
+      Set<String> stoppedReplicas = dataNodeConfig.getStoppedReplicas();
+      AmbryDataNode dataNode = instanceNameToAmbryDataNode.get(instanceName);
+      ConcurrentHashMap<String, AmbryReplica> currentReplicasOnNode = ambryDataNodeToAmbryReplicas.get(dataNode);
+      ConcurrentHashMap<String, AmbryReplica> replicasFromInstanceConfig = new ConcurrentHashMap<>();
+      Map<String, AmbryDisk> mountPathToDisk = ambryDataNodeToAmbryDisks.get(dataNode)
+          .stream()
+          .collect(Collectors.toMap(AmbryDisk::getMountPath, disk -> disk));
+      Map<AmbryPartition, List<AmbryReplica>> replicaToAddByPartition = new HashMap<>();
+      for (Map.Entry<String, DataNodeConfig.DiskConfig> diskEntry : dataNodeConfig.getDiskConfigs().entrySet()) {
+        String mountPath = diskEntry.getKey();
+        DataNodeConfig.DiskConfig diskConfig = diskEntry.getValue();
+        AmbryDisk disk = mountPathToDisk.getOrDefault(mountPath, null);
+        if (disk == null) {
+          logger.warn("{} is a new disk or unrecognizable disk which is not supported on existing node {}.", mountPath,
+              instanceName);
+          // TODO support dynamically adding disk in the future
+          continue;
+        }
+        // update disk capacity if needed
+        if (disk.getRawCapacityInBytes() != diskConfig.getDiskCapacityInBytes()) {
+          long prevDiskCapacity = disk.getRawCapacityInBytes();
+          logger.info("Capacity of disk at {} on {} has changed. Previous was: {} bytes, new capacity is {} bytes",
+              mountPath, instanceName, prevDiskCapacity, diskConfig.getDiskCapacityInBytes());
+          disk.setDiskCapacityInBytes(diskConfig.getDiskCapacityInBytes());
+          addClusterWideRawCapacity(diskConfig.getDiskCapacityInBytes() - prevDiskCapacity);
+        }
+        for (Map.Entry<String, DataNodeConfig.ReplicaConfig> replicaEntry : diskConfig.getReplicaConfigs().entrySet()) {
+          // partition name and replica name are the same.
+          String partitionName = replicaEntry.getKey();
+          DataNodeConfig.ReplicaConfig replicaConfig = replicaEntry.getValue();
+          if (currentReplicasOnNode.containsKey(partitionName)) {
+            // if replica is already present
+            AmbryReplica existingReplica = currentReplicasOnNode.get(partitionName);
+            // 1. directly add it into "replicasFromInstanceConfig" map
+            replicasFromInstanceConfig.put(partitionName, existingReplica);
+            // 2. update replica seal/stop state
+            updateReplicaStateAndOverrideIfNeeded(existingReplica, sealedReplicas, stoppedReplicas);
+          } else {
+            // if this is a new replica and doesn't exist on node
+            logger.info("Adding new replica {} to existing node {} in {}", partitionName, instanceName,
+                dcOrClusterName);
+            // this can be a brand new partition that is added to an existing node
+            AmbryPartition mappedPartition =
+                new AmbryPartition(Long.parseLong(partitionName), replicaConfig.getPartitionClass(),
+                    helixClusterManagerCallback);
+            // Ensure only one AmbryPartition instance exists for specific partition.
+            mappedPartition = addPartitionIfAbsent(mappedPartition, replicaConfig.getReplicaCapacityInBytes());
+            ensurePartitionAbsenceOnNodeAndValidateCapacity(mappedPartition, dataNode,
+                replicaConfig.getReplicaCapacityInBytes());
+            // create new replica belonging to this partition or find the existing replica from bootstrapReplicas map.
+            AmbryReplica replica;
+            if (selfInstanceName.equals(instanceName)) {
+              // if this is a newly added replica on current instance, it should be present in bootstrapReplicas map.
+              replica = fetchBootstrapReplica(mappedPartition.toPathString());
+              if (replica == null) {
+                logger.error("Replica {} is not present in bootstrap replica set, abort instance info update",
+                    mappedPartition.toPathString());
+                throw new IllegalStateException("Replica to add is not present in bootstrap replica map");
+              }
+            } else {
+              replica = new AmbryServerReplica(clusterMapConfig, mappedPartition, disk,
+                  stoppedReplicas.contains(partitionName), replicaConfig.getReplicaCapacityInBytes(),
+                  sealedReplicas.contains(partitionName));
+            }
+            updateReplicaStateAndOverrideIfNeeded(replica, sealedReplicas, stoppedReplicas);
+            // add new created replica to "replicasFromInstanceConfig" map
+            replicasFromInstanceConfig.put(partitionName, replica);
+            // Put new replica into partition-to-replica map temporarily (this is to avoid any exception thrown within the
+            // loop before updating "ambryDataNodeToAmbryReplicas" map. If we update call addReplicasToPartition here
+            // immediately, the exception may cause inconsistency between "ambryPartitionToAmbryReplicas" and
+            // "ambryDataNodeToAmbryReplicas")
+            replicaToAddByPartition.put(mappedPartition, Collections.singletonList(replica));
+          }
+        }
+      }
+      // update "ambryDataNodeToAmbryReplicas" map and "addedReplicas" list
+      replicaToAddByPartition.forEach((k, v) -> {
+        addReplicasToPartition(k, v);
+        addedReplicas.addAll(v);
       });
-      clusterWideAllocatedRawCapacityBytes.getAndAdd(replicas.get(0).getCapacityInBytes() * replicas.size());
+      // update ambryDataNodeToAmbryReplicas map by adding "replicasFromInstanceConfig"
+      ambryDataNodeToAmbryReplicas.put(instanceNameToAmbryDataNode.get(instanceName), replicasFromInstanceConfig);
+      // Derive old replicas that are removed and delete them from partition
+      currentReplicasOnNode.keySet()
+          .stream()
+          .filter(partitionName -> !replicasFromInstanceConfig.containsKey(partitionName))
+          .forEach(pName -> {
+            logger.info("Removing replica {} from existing node {}", pName, instanceName);
+            AmbryReplica ambryReplica = currentReplicasOnNode.get(pName);
+            removeReplicasFromPartition(ambryReplica.getPartitionId(), Collections.singletonList(ambryReplica));
+            removedReplicas.add(ambryReplica);
+          });
+      return new Pair<>(addedReplicas, removedReplicas);
     }
 
     /**
-     * Remove {@link AmbryReplica}(s) from given {@link AmbryPartition} and update cluster-wide capacity stats
-     * @param partition the {@link AmbryPartition} which replicas should be removed from.
-     * @param replicas list of {@link AmbryReplica} to be removed.
+     * If partition override is enabled, we override replica SEAL/UNSEAL state based on partitionOverrideMap. If disabled,
+     * update replica state according to the info from {@link DataNodeConfig}.
+     * @param replica the {@link ReplicaId} whose states (seal,stop) should be updated.
+     * @param sealedReplicas a collection of {@link ReplicaId}(s) that are in SEALED state.
+     * @param stoppedReplicas a collection of {@link ReplicaId}(s) that are in STOPPED state.
      */
-    void removeReplicasFromPartition(AmbryPartition partition, List<AmbryReplica> replicas) {
-      ambryPartitionToAmbryReplicas.computeIfPresent(partition, (k, v) -> {
-        v.removeAll(replicas);
-        clusterWideAllocatedRawCapacityBytes.getAndAdd(-1 * replicas.get(0).getCapacityInBytes() * replicas.size());
-        return v;
-      });
+    private void updateReplicaStateAndOverrideIfNeeded(AmbryReplica replica, Collection<String> sealedReplicas,
+        Collection<String> stoppedReplicas) {
+      String partitionName = replica.getPartitionId().toPathString();
+      boolean isSealed;
+      if (clusterMapConfig.clusterMapEnablePartitionOverride && partitionOverrideInfoMap.containsKey(partitionName)) {
+        isSealed = partitionOverrideInfoMap.get(partitionName)
+            .get(ClusterMapUtils.PARTITION_STATE)
+            .equals(ClusterMapUtils.READ_ONLY_STR);
+      } else {
+        isSealed = sealedReplicas.contains(partitionName);
+      }
+      replica.setSealedState(isSealed);
+      replica.setStoppedState(stoppedReplicas.contains(partitionName));
     }
 
     /**
-     * Add cluster-wide raw capacity. This is called when new disk is created and added to cluster. We update
-     * cluster-wide raw capacity by adding the disk capacity.
-     * @param diskRawCapacityBytes raw disk capacity to be added.
+     * Create a new instance(node) and initialize disks/replicas on it.
+     * @param dataNodeConfig the {@link DataNodeConfig} to create new instance
+     * @return a list of newly added replicas;
+     * @throws Exception if there is an exception in instantiating the {@link ResourceStatePolicy}
      */
-    void addClusterWideRawCapacity(long diskRawCapacityBytes) {
-      clusterWideRawCapacityBytes.getAndAdd(diskRawCapacityBytes);
+    private List<ReplicaId> createNewInstance(DataNodeConfig dataNodeConfig) throws Exception {
+      String instanceName = dataNodeConfig.getInstanceName();
+      logger.info("Adding node {} and its disks and replicas in {}", instanceName, dcOrClusterName);
+      AmbryDataNode datanode =
+          new AmbryServerDataNode(dataNodeConfig.getDatacenterName(), clusterMapConfig, dataNodeConfig.getHostName(),
+              dataNodeConfig.getPort(), dataNodeConfig.getRackId(), dataNodeConfig.getSslPort(),
+              dataNodeConfig.getHttp2Port(), DEFAULT_XID, helixClusterManagerCallback);
+      // for new instance, we first set it to unavailable and rely on its participation to update its liveness
+      if (!instanceName.equals(selfInstanceName)) {
+        datanode.setState(HardwareState.UNAVAILABLE);
+      }
+      List<ReplicaId> addedReplicas = initializeDisksAndReplicasOnNode(datanode, dataNodeConfig);
+      instanceNameToAmbryDataNode.put(instanceName, datanode);
+      dcToNodes.computeIfAbsent(datanode.getDatacenterName(), s -> new HashSet<>()).add(datanode);
+      allInstances.add(instanceName);
+      return addedReplicas;
     }
 
     /**
-     * Pop out bootstrap replica (if any) on current instance. A bootstrap replica is a replica dynamically added to
-     * current node at runtime.
-     * @param partitionName the partition name of bootstrap replica.
-     * @return bootstrap replica or {@code null} if not found.
+     * Initialize the disks and replicas on the given node. Create partitions if this is the first time a replica of
+     * that partition is being constructed. If partition override is enabled, the seal state of replica is determined by
+     * partition info in HelixPropertyStore, if disabled, the seal state is determined by {@code dataNodeConfig}.
+     * @param datanode the {@link AmbryDataNode} that is being initialized.
+     * @param dataNodeConfig the {@link DataNodeConfig} associated with this datanode.
+     * @return a list of newly added replicas on this node.
+     * @throws Exception if creation of {@link AmbryDisk} throws an Exception.
      */
-    AmbryReplica fetchBootstrapReplica(String partitionName) {
-      return (AmbryReplica) bootstrapReplicas.remove(partitionName);
+    private List<ReplicaId> initializeDisksAndReplicasOnNode(AmbryDataNode datanode, DataNodeConfig dataNodeConfig)
+        throws Exception {
+      List<ReplicaId> addedReplicas = new ArrayList<>();
+      Set<String> sealedReplicas = dataNodeConfig.getSealedReplicas();
+      Set<String> stoppedReplicas = dataNodeConfig.getStoppedReplicas();
+      ambryDataNodeToAmbryReplicas.put(datanode, new ConcurrentHashMap<>());
+      ambryDataNodeToAmbryDisks.put(datanode, new HashSet<>());
+      for (Map.Entry<String, DataNodeConfig.DiskConfig> diskEntry : dataNodeConfig.getDiskConfigs().entrySet()) {
+        String mountPath = diskEntry.getKey();
+        DataNodeConfig.DiskConfig diskConfig = diskEntry.getValue();
+
+        // Create disk
+        AmbryDisk disk = new AmbryDisk(clusterMapConfig, datanode, mountPath, diskConfig.getState(),
+            diskConfig.getDiskCapacityInBytes());
+        ambryDataNodeToAmbryDisks.get(datanode).add(disk);
+        addClusterWideRawCapacity(diskConfig.getDiskCapacityInBytes());
+
+        for (Map.Entry<String, DataNodeConfig.ReplicaConfig> replicaEntry : diskConfig.getReplicaConfigs().entrySet()) {
+          String partitionName = replicaEntry.getKey();
+          DataNodeConfig.ReplicaConfig replicaConfig = replicaEntry.getValue();
+
+          AmbryPartition mappedPartition =
+              new AmbryPartition(Long.parseLong(partitionName), replicaConfig.getPartitionClass(),
+                  helixClusterManagerCallback);
+          // Ensure only one AmbryPartition instance exists for specific partition.
+          mappedPartition = addPartitionIfAbsent(mappedPartition, replicaConfig.getReplicaCapacityInBytes());
+          ensurePartitionAbsenceOnNodeAndValidateCapacity(mappedPartition, datanode,
+              replicaConfig.getReplicaCapacityInBytes());
+          // Create replica associated with this node and this partition
+          boolean isSealed = sealedReplicas.contains(partitionName);
+          if (clusterMapConfig.clusterMapEnablePartitionOverride && partitionOverrideInfoMap.containsKey(
+              partitionName)) {
+            // override sealed state if PartitionOverride is enabled.
+            isSealed = partitionOverrideInfoMap.get(partitionName)
+                .get(ClusterMapUtils.PARTITION_STATE)
+                .equals(ClusterMapUtils.READ_ONLY_STR);
+          }
+          AmbryReplica replica =
+              new AmbryServerReplica(clusterMapConfig, mappedPartition, disk, stoppedReplicas.contains(partitionName),
+                  replicaConfig.getReplicaCapacityInBytes(), isSealed);
+          ambryDataNodeToAmbryReplicas.get(datanode).put(mappedPartition.toPathString(), replica);
+          addReplicasToPartition(mappedPartition, Collections.singletonList(replica));
+          addedReplicas.add(replica);
+        }
+      }
+      return addedReplicas;
+    }
+
+    /**
+     * Update the liveness states of existing instances based on the input.
+     * @param liveInstances the list of instances that are up.
+     */
+    private void updateInstanceLiveness(List<LiveInstance> liveInstances) {
+      Set<String> liveInstancesSet = new HashSet<>();
+      liveInstances.forEach(e -> liveInstancesSet.add(e.getInstanceName()));
+      for (String instanceName : allInstances) {
+        // Here we ignore live instance change it's about self instance. The reason is, during server's startup, current
+        // node should be AVAILABLE but the list of live instances doesn't include current node since it hasn't joined yet.
+        if (liveInstancesSet.contains(instanceName) || instanceName.equals(selfInstanceName)) {
+          instanceNameToAmbryDataNode.get(instanceName).setState(HardwareState.AVAILABLE);
+        } else {
+          instanceNameToAmbryDataNode.get(instanceName).setState(HardwareState.UNAVAILABLE);
+        }
+      }
+    }
+
+    /**
+     * Ensure that the given partition is absent on the given datanode. This is called as part of an inline validation
+     * done to ensure that two replicas of the same partition do not exist on the same datanode.
+     * @param partition the {@link AmbryPartition} to check.
+     * @param datanode the {@link AmbryDataNode} on which to check.
+     * @param expectedReplicaCapacity the capacity expected for the replicas of the partition.
+     */
+    private void ensurePartitionAbsenceOnNodeAndValidateCapacity(AmbryPartition partition, AmbryDataNode datanode,
+        long expectedReplicaCapacity) {
+      for (AmbryReplica replica : helixClusterManagerCallback.getReplicaIdsForPartition(partition)) {
+        if (replica.getDataNodeId().equals(datanode)) {
+          throw new IllegalStateException("Replica already exists on " + datanode + " for " + partition);
+        } else if (replica.getCapacityInBytes() != expectedReplicaCapacity) {
+          throw new IllegalStateException("Expected replica capacity " + expectedReplicaCapacity + " is different from "
+              + "the capacity of an existing replica " + replica.getCapacityInBytes());
+        }
+      }
     }
   }
 
   /**
-   * A callback class used to query information from the {@link HelixClusterManager}
+   * A callback class used to update and query cluster information from the {@link HelixClusterManager}
    */
   class HelixClusterManagerCallback
       implements ClusterManagerCallback<AmbryReplica, AmbryDisk, AmbryPartition, AmbryDataNode> {
@@ -774,12 +1136,9 @@ public class HelixClusterManager implements ClusterMap {
     @Override
     public String getResourceNameForPartition(AmbryPartition partition) {
       String result = null;
-      if (clusterMapConfig.clusterMapUseAggregatedView) {
-        return clusterInfo.clusterChangeHandler.getPartitionToResourceMap().get(partition.toPathString());
-      }
       // go through all datacenters in case that partition resides in one datacenter only
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        String resourceName = dcInfo.clusterChangeHandler.getPartitionToResourceMap().get(partition.toPathString());
+      for (String dcName : partitionToResourceNameByDc.keySet()) {
+        String resourceName = partitionToResourceNameByDc.get(dcName).get(partition.toPathString());
         if (resourceName != null) {
           result = resourceName;
           break;
@@ -790,23 +1149,39 @@ public class HelixClusterManager implements ClusterMap {
 
     /**
      * {@inheritDoc}
-     * If dcName is null, then get replicas by given state from all datacenters.
+     * If dcOrClusterName is null, then get replicas by given state from all datacenters.
      * If no routing table snapshot is found for dc name, or no resource name found for given partition, return empty list.
      */
     @Override
     public List<AmbryReplica> getReplicaIdsByState(AmbryPartition partition, ReplicaState state, String dcName) {
       List<AmbryReplica> replicas = new ArrayList<>();
       if (clusterMapConfig.clusterMapUseAggregatedView) {
-        clusterInfo.clusterChangeHandler.getReplicaIdsByState(partition, state).forEach(ambryReplica -> {
-          if (dcName == null || ambryReplica.getDataNodeId().getDatacenterName().equals(dcName)) {
-            replicas.add(ambryReplica);
-          }
-        });
+        String resourceName = dcName != null ? partitionToResourceNameByDc.get(dcName).get(partition.toPathString())
+            : getResourceNameForPartition(partition);
+        globalRoutingTableSnapshotRef.get()
+            .getInstancesForResource(resourceName, partition.toPathString(), state.name())
+            .stream()
+            .map(instanceConfig -> instanceNameToAmbryDataNode.get(instanceConfig.getInstanceName()))
+            .map(dataNode -> ambryDataNodeToAmbryReplicas.get(dataNode).get(partition.toPathString()))
+            .filter(Objects::nonNull)
+            .forEach(replica -> {
+              if (dcName == null || replica.getDataNodeId().getDatacenterName().equals(dcName)) {
+                replicas.add(replica);
+              }
+            });
       } else {
         for (DcInfo dcInfo : dcToDcInfo.values()) {
           String dc = dcInfo.dcName;
           if (dcName == null || dcName.equals(dc)) {
-            dcInfo.clusterChangeHandler.getReplicaIdsByState(partition, state).forEach(replicas::add);
+            String resourceName = partitionToResourceNameByDc.get(dc).get(partition.toPathString());
+            dcToRoutingTableSnapshotRef.get(dcInfo.dcName)
+                .get()
+                .getInstancesForResource(resourceName, partition.toPathString(), state.name())
+                .stream()
+                .map(instanceConfig -> instanceNameToAmbryDataNode.get(instanceConfig.getInstanceName()))
+                .map(dataNode -> ambryDataNodeToAmbryReplicas.get(dataNode).get(partition.toPathString()))
+                .filter(Objects::nonNull)
+                .forEach(replicas::add);
           }
         }
       }
@@ -817,13 +1192,37 @@ public class HelixClusterManager implements ClusterMap {
     public void getReplicaIdsByStates(Map<ReplicaState, List<AmbryReplica>> replicasByState, AmbryPartition partition,
         Set<ReplicaState> states, String dcName) {
       if (clusterMapConfig.clusterMapUseAggregatedView) {
-        clusterInfo.clusterChangeHandler.getReplicaIdsByStates(replicasByState, states, partition);
-
+        String resourceName = dcName != null ? partitionToResourceNameByDc.get(dcName).get(partition.toPathString())
+            : getResourceNameForPartition(partition);
+        for (ReplicaState state : states) {
+          List<AmbryReplica> list = replicasByState.computeIfAbsent(state, k -> new ArrayList<>());
+          globalRoutingTableSnapshotRef.get()
+              .getInstancesForResource(resourceName, partition.toPathString(), state.name())
+              .stream()
+              .map(instanceConfig -> instanceNameToAmbryDataNode.get(instanceConfig.getInstanceName()))
+              .map(dataNode -> ambryDataNodeToAmbryReplicas.get(dataNode).get(partition.toPathString()))
+              .filter(Objects::nonNull)
+              .forEach(replica -> {
+                if (dcName == null || replica.getDataNodeId().getDatacenterName().equals(dcName)) {
+                  list.add(replica);
+                }
+              });
+        }
       } else {
         for (DcInfo dcInfo : dcToDcInfo.values()) {
           String dc = dcInfo.dcName;
           if (dcName == null || dcName.equals(dc)) {
-            dcInfo.clusterChangeHandler.getReplicaIdsByStates(replicasByState, states, partition);
+            String resourceName = partitionToResourceNameByDc.get(dc).get(partition.toPathString());
+            RoutingTableSnapshot snapshot = dcToRoutingTableSnapshotRef.get(dc).get();
+            for (ReplicaState state : states) {
+              List<AmbryReplica> list = replicasByState.computeIfAbsent(state, k -> new ArrayList<>());
+              snapshot.getInstancesForResource(resourceName, partition.toPathString(), state.name())
+                  .stream()
+                  .map(instanceConfig -> instanceNameToAmbryDataNode.get(instanceConfig.getInstanceName()))
+                  .map(dataNode -> ambryDataNodeToAmbryReplicas.get(dataNode).get(partition.toPathString()))
+                  .filter(Objects::nonNull)
+                  .forEach(list::add);
+            }
           }
         }
       }
@@ -875,15 +1274,14 @@ public class HelixClusterManager implements ClusterMap {
     @Override
     public Collection<AmbryDisk> getDisks(AmbryDataNode dataNode) {
       if (dataNode != null) {
-        return dcToDcInfo.get(dataNode.getDatacenterName()).clusterChangeHandler.getDisks(dataNode);
-      }
-      List<AmbryDisk> disksToReturn = new ArrayList<>();
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        for (Set<AmbryDisk> disks : dcInfo.clusterChangeHandler.getDataNodeToDisksMap().values()) {
+        return ambryDataNodeToAmbryDisks.get(dataNode);
+      } else {
+        List<AmbryDisk> disksToReturn = new ArrayList<>();
+        for (Set<AmbryDisk> disks : ambryDataNodeToAmbryDisks.values()) {
           disksToReturn.addAll(disks);
         }
+        return disksToReturn;
       }
-      return disksToReturn;
     }
 
     /**
@@ -891,10 +1289,8 @@ public class HelixClusterManager implements ClusterMap {
      */
     long getDiskCount() {
       long count = 0;
-      for (DcInfo dcInfo : dcToDcInfo.values()) {
-        for (Set<AmbryDisk> disks : dcInfo.clusterChangeHandler.getDataNodeToDisksMap().values()) {
-          count += disks.size();
-        }
+      for (Set<AmbryDisk> disks : ambryDataNodeToAmbryDisks.values()) {
+        count += disks.size();
       }
       return count;
     }
@@ -904,9 +1300,11 @@ public class HelixClusterManager implements ClusterMap {
      */
     long getDownDisksCount() {
       long count = 0;
-      for (AmbryDisk disk : getDisks(null)) {
-        if (disk.getState() == HardwareState.UNAVAILABLE) {
-          count++;
+      for (Set<AmbryDisk> disks : ambryDataNodeToAmbryDisks.values()) {
+        for (AmbryDisk disk : disks) {
+          if (disk.getState() == HardwareState.UNAVAILABLE) {
+            count++;
+          }
         }
       }
       return count;
@@ -972,6 +1370,17 @@ public class HelixClusterManager implements ClusterMap {
      */
     long getAllocatedUsableCapacity() {
       return clusterWideAllocatedUsableCapacityBytes.get();
+    }
+
+    /**
+     * Gets the {@link RoutingTableSnapshot} for a given data center or entire cluster.
+     * @param dcName data center name.
+     * @return {@link RoutingTableSnapshot} for a given data center. If the input data center name is null or if using
+     * aggregated view, returns the routing table snapshot of entire cluster.
+     */
+    public RoutingTableSnapshot getRoutingTableSnapshot(String dcName) {
+      return dcName == null || clusterMapConfig.clusterMapUseAggregatedView ? globalRoutingTableSnapshotRef.get()
+          : dcToRoutingTableSnapshotRef.get(dcName).get();
     }
   }
 }
