+++ ambry-clustermap/src/test/java/com/github/ambry/clustermap/HelixClusterManagerTest.java
@@ -49,6 +49,7 @@ import org.apache.helix.zookeeper.datamodel.ZNRecord;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.After;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -83,8 +84,6 @@ public class HelixClusterManagerTest {
   private final String selfInstanceName;
   private final String localDc;
   private final String remoteDc;
-  private final String cloudDc;
-  private final int numCloudDcs;
   private ClusterMap clusterManager;
   private MetricRegistry metricRegistry;
   private Map<String, Gauge> gauges;
@@ -131,14 +130,6 @@ public class HelixClusterManagerTest {
     MockitoAnnotations.initMocks(this);
     localDc = helixDcs[0];
     remoteDc = helixDcs[1];
-    if (!useComposite) {
-      // cloud DCs only supported in HelixClusterManager
-      cloudDc = "cloud-dc";
-      numCloudDcs = 1;
-    } else {
-      cloudDc = null;
-      numCloudDcs = 0;
-    }
     Random random = new Random();
     File tempDir = Files.createTempDirectory("helixClusterManager-" + random.nextInt(1000)).toFile();
     String tempDirPath = tempDir.getAbsolutePath();
@@ -152,9 +143,6 @@ public class HelixClusterManagerTest {
     partitionLayoutPath = tempDirPath + File.separator + "partitionLayoutTest.json";
     String zkLayoutPath = tempDirPath + File.separator + "zkLayoutPath.json";
     zkJson = constructZkLayoutJSON(dcsToZkInfo.values());
-    if (!useComposite) {
-      addCloudDc(zkJson, dcId, cloudDc);
-    }
     // initial partition count = 3
     testHardwareLayout = constructInitialHardwareLayoutJSON(clusterNameStatic);
     testPartitionLayout = constructInitialPartitionLayoutJSON(testHardwareLayout, 3, localDc);
@@ -345,9 +333,6 @@ public class HelixClusterManagerTest {
     dcsToZkInfo.get(remoteDc).setPort(0);
     Set<com.github.ambry.utils.TestUtils.ZkInfo> zkInfos = new HashSet<>(dcsToZkInfo.values());
     JSONObject invalidZkJson = constructZkLayoutJSON(zkInfos);
-    if (cloudDc != null) {
-      addCloudDc(invalidZkJson, (byte) zkInfos.size(), cloudDc);
-    }
     Properties props = new Properties();
     props.setProperty("clustermap.host.name", hostname);
     props.setProperty("clustermap.port", Integer.toString(portNum));
@@ -395,33 +380,6 @@ public class HelixClusterManagerTest {
           metricRegistry.getGauges().get(HelixClusterManager.class.getName() + ".instantiationFailed").getValue());
       assertEquals("beBad", e.getCause().getMessage());
     }
-  }
-
-  /**
-   * Test startup when setting the cloud datacenter as the local datacenter.
-   * @throws Exception
-   */
-  @Test
-  public void testCloudDcAsLocal() throws Exception {
-    assumeTrue(cloudDc != null);
-    assumeTrue(!overrideEnabled);
-
-    String hostname = "localhost";
-    String selfInstanceName = getInstanceName(hostname, null);
-    Properties props = new Properties();
-    props.setProperty("clustermap.host.name", hostname);
-    props.setProperty("clustermap.cluster.name", clusterNamePrefixInHelix + clusterNameStatic);
-    props.setProperty("clustermap.datacenter.name", cloudDc);
-    props.setProperty("clustermap.dcs.zk.connect.strings", zkJson.toString(2));
-    props.setProperty("clustermap.current.xid", Long.toString(CURRENT_XID));
-    props.setProperty("clustermap.enable.partition.override", Boolean.toString(overrideEnabled));
-    props.setProperty("clustermap.listen.cross.colo", Boolean.toString(listenCrossColo));
-    ClusterMapConfig clusterMapConfig = new ClusterMapConfig(new VerifiableProperties(props));
-    MockHelixManagerFactory helixManagerFactory = new MockHelixManagerFactory(helixCluster, new HashMap<>(), null);
-    metricRegistry = new MetricRegistry();
-    clusterManager = new HelixClusterManager(clusterMapConfig, selfInstanceName, helixManagerFactory, metricRegistry);
-
-    verifyInitialClusterChanges((HelixClusterManager) clusterManager, helixCluster, helixDcs);
   }
 
   /**
@@ -1018,7 +976,7 @@ public class HelixClusterManagerTest {
         aliveCount++;
       }
     }
-    assertEquals("Mismatch in number of alive replicas", instances.size() + numCloudDcs - 1, aliveCount);
+    assertEquals("Mismatch in number of alive replicas", instances.size() - 1, aliveCount);
     // unmark the stopped replica and no replica is in stopped state
     helixCluster.setReplicaState(partition, instances.get(0), ReplicaStateType.StoppedState, false, false);
     aliveCount = 0;
@@ -1027,8 +985,7 @@ public class HelixClusterManagerTest {
         aliveCount++;
       }
     }
-    assertEquals("Mismatch in number of alive replicas, all replicas should be up", instances.size() + numCloudDcs,
-        aliveCount);
+    assertEquals("Mismatch in number of alive replicas, all replicas should be up", instances.size(), aliveCount);
   }
 
   /**
@@ -1050,7 +1007,9 @@ public class HelixClusterManagerTest {
   /**
    * Tests that if the xid of an InstanceConfig change is greater than the current xid of the cluster manager, then that
    * change is ignored - both during initialization as well as with post-initialization InstanceConfig changes.
+   * TODO: xid is no longer used. This test can be removed after removing xid in rest of places.
    */
+  @Ignore
   @Test
   public void xidTest() throws Exception {
     assumeTrue(!useComposite && listenCrossColo);
@@ -1075,7 +1034,7 @@ public class HelixClusterManagerTest {
     aheadInstanceConfig.getRecord().setSimpleField(XID_STR, Long.toString(CURRENT_XID + 1));
     clusterManager =
         new HelixClusterManager(clusterMapConfig, selfInstanceName, helixManagerFactory, new MetricRegistry());
-    assertEquals(instanceCount + numCloudDcs - 1, clusterManager.getDataNodeIds().size());
+    assertEquals(instanceCount - 1, clusterManager.getDataNodeIds().size());
     for (DataNodeId dataNode : clusterManager.getDataNodeIds()) {
       String instanceName = ClusterMapUtils.getInstanceName(dataNode.getHostname(), dataNode.getPort());
       assertThat(instanceName, not(aheadInstanceConfig.getInstanceName()));
@@ -1084,7 +1043,7 @@ public class HelixClusterManagerTest {
     // Ahead instance should be honored if the cluster manager is of the aheadInstance.
     try (HelixClusterManager aheadInstanceClusterManager = new HelixClusterManager(clusterMapConfig,
         aheadInstanceConfig.getInstanceName(), helixManagerFactory, new MetricRegistry())) {
-      assertEquals(instanceCount + numCloudDcs, aheadInstanceClusterManager.getDataNodeIds().size());
+      assertEquals(instanceCount, aheadInstanceClusterManager.getDataNodeIds().size());
     }
 
     // Post-initialization InstanceConfig change: pick an instance that is neither previous instance nor self instance
@@ -1192,7 +1151,7 @@ public class HelixClusterManagerTest {
     // node brought up in each DC.
     assertEquals(instanceTriggerCount, getCounterValue("liveInstanceChangeTriggerCount"));
     assertEquals(helixDcs.length, getCounterValue("instanceConfigChangeTriggerCount"));
-    assertEquals(helixCluster.getDataCenterCount() + numCloudDcs, getGaugeValue("datacenterCount"));
+    assertEquals(helixCluster.getDataCenterCount(), getGaugeValue("datacenterCount"));
     assertEquals(helixCluster.getDownInstances().size() + helixCluster.getUpInstances().size(),
         getGaugeValue("dataNodeCount"));
     assertEquals(helixCluster.getDownInstances().size(), getGaugeValue("dataNodeDownCount"));
@@ -1301,19 +1260,6 @@ public class HelixClusterManagerTest {
         }
       }
     }
-    if (cloudDc != null) {
-      // If one cloud DC is present, there should be exactly one virtual replica for every partition.
-      for (PartitionId partitionId : clusterManager.getAllPartitionIds(null)) {
-        List<? extends ReplicaId> replicaIds = partitionId.getReplicaIds();
-        int count = 0;
-        for (ReplicaId replicaId : replicaIds) {
-          if (replicaId instanceof CloudServiceReplica) {
-            count++;
-          }
-        }
-        assertEquals("Unexpected number of CloudServiceReplicas in partition: " + replicaIds, 1, count);
-      }
-    }
   }
 
   /**
@@ -1391,7 +1337,7 @@ public class HelixClusterManagerTest {
    */
   private void testPartitionReplicaConsistency() throws Exception {
     for (PartitionId partition : clusterManager.getWritablePartitionIds(null)) {
-      assertEquals(testPartitionLayout.getTotalReplicaCount() + numCloudDcs, partition.getReplicaIds().size());
+      assertEquals(testPartitionLayout.getTotalReplicaCount(), partition.getReplicaIds().size());
       InputStream partitionStream = new ByteBufferInputStream(ByteBuffer.wrap(partition.getBytes()));
       PartitionId fetchedPartition = clusterManager.getPartitionIdFromStream(partitionStream);
       assertEquals(partition, fetchedPartition);
@@ -1465,17 +1411,6 @@ public class HelixClusterManagerTest {
     Pair<Set<String>, Set<String>> writablePartitionsInTwoPlaces = getWritablePartitions();
     assertEquals(writablePartitionsInTwoPlaces.getFirst(), writablePartitionsInTwoPlaces.getSecond());
     testAllPartitions();
-  }
-
-  /**
-   * @param zkJson the JSON to add the datacenter info to.
-   * @param dcId the datacenter ID.
-   * @param cloudDc the dc name
-   */
-  private static void addCloudDc(JSONObject zkJson, byte dcId, String cloudDc) {
-    zkJson.append(ZKINFO_STR, new JSONObject().put(DATACENTER_STR, cloudDc)
-        .put(DATACENTER_ID_STR, dcId)
-        .put(REPLICA_TYPE_STR, ReplicaType.CLOUD_BACKED));
   }
 
   /**
