+++ config/azureconfig/vcr.properties
@@ -30,4 +30,3 @@ vcr.cluster.zk.connect.string=1.linkedin.com:12914
 vcr.cluster.name=AmbrDev-VCR
 clustermap.dcs.zk.connect.strings={"zkInfo":[{"datacenter":"ei1","id":1,"zkConnectStr":"1.linkedin.com:12914"}]}
 vcr.source.datacenters=ei1
-clustermap.cluster.change.handler.type=DynamicClusterChangeHandler
